Implement update and delete.
Changed the model for choices so that deletes cascade. However, upon
doing this, I realized that sequelize won't alter an already existing
table with a different rule for deletions to change the deletion rule.

So I implemented a different deletion method that does not assume
cascading.

Changes to be committed:
      modified:   controllers/polls.js
      modified:   models/choices.js